requirements.txt
praw==7.6.0
python-telegram-bot==13.13

